refactor actions

+++ src/actions/address/delete-user.address.ts
@@ -4,11 +4,24 @@ import prisma from '@/lib/prisma';
 
 export const deleteUserAddress = async (userId: string) => {
   try {
+    const userAddress = await prisma.userAddress.findUnique({
+      where: {
+        userId,
+      },
+    });
+
+    if (!userAddress) {
+      return {
+        ok: false,
+        message: 'No se encontro la dirección',
+      };
+    }
     const deletedAddress = await prisma.userAddress.delete({
       where: {
         userId,
       },
     });
+
     return {
       ok: true,
       deletedAddress,
